requirements.txt
pyyaml>=6.0.1
requests>=2.31.0
fastapi>=0.111.0
uvicorn>=0.23.0
pydantic>=2.6.0